login,signup succesfull & all routes protected

+++ frontend/src/componets/OpenCourseDetails.jsx
@@ -1,8 +1,18 @@
-const  OpenCourseDetails=({chapter})=>{
+import { useAuthContext } from "../hooks/useAuthContext"
 
+const  OpenCourseDetails=({chapter})=>{
+    
+    const {user}=useAuthContext()
     const handleClick= async()=>{
+        if(!user){
+            return
+        }
+
         const response=await fetch('http://localhost:4000/api/chapters/'+ chapter._id,{
-            method:'DELETE'
+            method:'DELETE',
+            headers:{
+                'Authorization':`Bearer ${user.token}`
+            }
         })
         const json=await response.json()
